refactor: #332 HomeUseCase 프로토콜 분리
